pyproject.toml
[tool.poetry]
name = "pic2cal"
version = "0.1.0"
description = ""
authors = ["Leonid Sokolov <leonid.sokolov@big-picture.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.4"
openai = "^1.35.13"
pytest = "^8.2.2"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
